chatbot_backend/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.43.1
anyio==4.8.0
asgiref==3.8.1
async-timeout==5.0.1
attrs==24.3.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
defusedxml==0.8.0rc2
distro==1.9.0
dj-rest-auth==7.0.1
Django==4.2.18
django-allauth==65.3.1
django-cors-headers==4.6.0
djangorestframework==3.15.2
ecdsa==0.19.0
exceptiongroup==1.2.2
frozenlist==1.5.0
groq==0.15.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
multidict==6.1.0
oauthlib==3.2.2
openai==1.54.0
pillow==11.1.0
propcache==0.2.1
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
PyJWT==2.10.1
python-jose==3.3.0
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.17.0
sniffio==1.3.1
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
yarl==1.18.3

python-dotenv==1.1.1
ollama==0.6.0
opencv-python==4.10.0.84 
PyPDF2==3.0.1
pdf2image==1.17.0
pytesseract==0.3.13

yolov5
torch
torchvision
torchaudio

# 🔥 하이브리드 분석용 라이브러리
deepface==0.0.92
transformers==4.36.2
tf-keras==2.16.0
opencv-python-headless==4.8.1.78
